+++ crates/teloxide-core/src/types/chat_administrator_rights.rs
@@ -47,6 +47,18 @@ pub struct ChatAdministratorRights {
     /// supergroups only
     pub can_pin_messages: Option<bool>,
 
+    /// `true`, if the administrator can post stories in the channel;
+    /// channels only
+    pub can_post_stories: Option<bool>,
+
+    /// `true`, if the administrator can edit stories posted by other users;
+    /// channels only
+    pub can_edit_stories: Option<bool>,
+
+    /// `true`, if the administrator can delete stories posted by other users;
+    /// channels only
+    pub can_delete_stories: Option<bool>,
+
     /// `true`, if the user is allowed to create, rename, close, and reopen
     /// forum topics; supergroups only
     pub can_manage_topics: Option<bool>,