+++ crates/teloxide-core/src/types/write_access_allowed.rs
@@ -10,4 +10,10 @@ use serde::{Deserialize, Serialize};
 pub struct WriteAccessAllowed {
     /// Name of the Web App which was launched from a link
     pub web_app_name: Option<String>,
+    /// `true`, if the access was granted after the user accepted an explicit
+    /// request from a Web App sent by the method [requestWriteAccess](https://core.telegram.org/bots/webapps#initializing-mini-apps)
+    pub from_request: Option<bool>,
+    /// `true`, if the access was granted when the bot was added to the
+    /// attachment or side menu
+    pub from_attachment_menu: Option<bool>,
 }